generate fake devtest output

+++ src/content_selection.py
@@ -7,16 +7,17 @@ from nltk import pos_tag
 from tqdm import tqdm
 import random
 import pprint
+from datetime import datetime
 
 
-def make_placeholder_summaries(input_data, data_store):
+def make_placeholder_summaries(input_data, data_store, unique_alphanum):
     for topic_id in tqdm(input_data.keys()):
         for doc_id, doc in input_data[topic_id]["docs"].items():
-            _create_placeholders(topic_id, doc_id, doc, out_dir=data_store["output_dir"])
+            _create_placeholders(topic_id, doc_id, doc, unique_alphanum, out_dir=data_store["output_dir"])
             break
 
 
-def make_placeholder_output(input_data, data_store):
+def make_placeholder_output(input_data, data_store, unique_alphanum):
     """
     Returns:
         A list of dicts, where each dict represents a sentence. 
@@ -32,16 +33,18 @@ def make_placeholder_output(input_data, data_store):
     all_data = []
     for topic_id in tqdm(input_data.keys()):
         for doc_id, doc in input_data[topic_id]["docs"].items():
-            doc_data = _create_placeholders(topic_id, doc_id, doc)
+            doc_data = _create_placeholders(
+                topic_id, doc_id, doc, unique_alphanum)
             all_data.extend(doc_data)
     return all_data
 
 
-def _create_placeholders(topic_id, doc_id, doc, unique_alphanum="D2run0", out_dir=None):
+def _create_placeholders(topic_id, doc_id, doc, unique_alphanum, out_dir=None, overwrite=True):
     random.seed(100)
     n_subtopics = 5 
 
     sentences = sent_tokenize(doc)
+    sentences = sentences[2:] # Skip the topic and date line
     summary_tokens = []
     summary_lines = []
     sentence_data = []
@@ -72,10 +75,14 @@ def _create_placeholders(topic_id, doc_id, doc, unique_alphanum="D2run0", out_di
             id_part_2=id_part_2,
             unique_alphanum=unique_alphanum,
         )
-        output_path = os.path.join(output_dir, output_name)
-        with open(output_path, "w") as outfile:
-            for line in summary_lines:
-                outfile.write(line + "\n")
+        output_path = os.path.join(out_dir, output_name)
+        if not os.path.exists(out_dir):
+            os.makedirs(out_dir)
+
+        if not os.path.exists(output_path) or overwrite:
+            with open(output_path, "w") as outfile:
+                for line in summary_lines:
+                    outfile.write(line + "\n")
     
     return sentence_data
 
@@ -84,6 +91,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--config", type=str, default="config.json")
     parser.add_argument("--split", type=str, default="training", choices=["devtest", "evaltest", "training"])
+    parser.add_argument("--deliverable", type=str, default="D2")
     args = parser.parse_args() 
 
     with open(args.config) as infile:
@@ -92,7 +100,12 @@ if __name__ == "__main__":
     if not os.path.exists(data_store["working_dir"]):
         os.makedirs(data_store["working_dir"])
 
-    input_data = load_data("input_data", data_store, "training", year=2009)
-    placeholder_output = make_placeholder_output(input_data, data_store)
-    for sent in placeholder_output[:5]:
+    input_data = load_data("input_data", data_store, "devtest", year=2010, test=True)
+
+    # run_id = args.deliverable + datetime.now().strftime('%Y%m%d%H%M%S')
+    run_id = "D2run0"
+    make_placeholder_summaries(input_data, data_store, run_id)
+    placeholder_output = make_placeholder_output(
+        input_data, data_store, run_id)
+    for sent in placeholder_output[:3]:
         pprint.pprint(sent)
